# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Solvro Backend",
      "version": "1.0.0"
    },
    "paths": {
...
      "/path/{source}/{target}": {
        "get": {
          "summary": "Returns list of stops in path and total distance.",
          "parameters": [
            {
              "in": "path",
              "name": "source",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "Stop where the path begins"
            },
            {
              "in": "path",
              "name": "target",
              "required": true,
              "schema": {
                "type": "string"
              },
              "description": "Stop where the path ends"
            }
          ],
          "responses": {
            "200": {
              "description": "JSON containing status, stops and total distance",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object"
                  },
                  "example": {
                    "status": true,
                    "from": 1,
                    "to": 3,
                    "path": {
                      "distance": 13,
                      "stops": [
                        {
                          "name": "Przystanek 1"
                        },
                        {
                          "name": "Przystanek 2"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/secured": {
        "get": {
          "summary": "Route that can only be accessed by logged-in users",
          "responses": {
            "200": {
              "description": "JSON containing 200 status code if logged."
            },
            "401": {
              "description": "Not logged in."
            }
          }
        }
      },
...
    }
  }